requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
charset-normalizer==2.1.0
discord-py-interactions==5.6.0
discord-typings==0.6.0
emoji==2.2.0
Flask==2.3.2
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.2
python-dotenv==0.20.0
typing_extensions==4.3.0
Werkzeug==2.3.6
yarl==1.7.2
